lots of minor updates (do not remember) and added CPUID

+++ utils/vutils.cpp
@@ -157,21 +157,21 @@ void simd_print(const char * const str, const SIMD_INT va)
     int itmp[SIMD_STREAMS_32] __SIMD_ALIGN__;
     simd_store(itmp, va);
     printf("%s", str);
-    for (int i = 0; i < SIMD_STREAMS_32; ++i)
+    for (size_t i = 0; i < SIMD_STREAMS_32; ++i)
       printf("%d\t", itmp[i]);
     printf("\n");
 
     unsigned int utmp[SIMD_STREAMS_32] __SIMD_ALIGN__;
     simd_store(utmp, va);
     printf("%s", str);
-    for (int i = 0; i < SIMD_STREAMS_32; ++i)
+    for (size_t i = 0; i < SIMD_STREAMS_32; ++i)
       printf("%u\t", utmp[i]);
     printf("\n");
 
     unsigned long int lutmp[SIMD_STREAMS_64] __SIMD_ALIGN__;
     simd_store(lutmp, va);
     printf("%s", str);
-    for (int i = 0; i < SIMD_STREAMS_64; ++i)
+    for (size_t i = 0; i < SIMD_STREAMS_64; ++i)
       printf("%lu\t", lutmp[i]);
     printf("\n");
 }
@@ -182,7 +182,7 @@ void simd_print(const char * const str, const SIMD_FLT va)
     float tmp[SIMD_STREAMS_32] __SIMD_ALIGN__;
     simd_store(tmp, va);
     printf("%s", str);
-    for (int i = 0; i < SIMD_STREAMS_32; ++i)
+    for (size_t i = 0; i < SIMD_STREAMS_32; ++i)
       printf("%f\t", tmp[i]);
     printf("\n");
 }
@@ -193,11 +193,10 @@ void simd_print(const char * const str, const SIMD_DBL va)
     double tmp[SIMD_STREAMS_64] __SIMD_ALIGN__;
     simd_store(tmp, va);
     printf("%s", str);
-    for (int i = 0; i < SIMD_STREAMS_64; ++i)
+    for (size_t i = 0; i < SIMD_STREAMS_64; ++i)
       printf("%f\t", tmp[i]);
     printf("\n");
 }
 
 
 #endif // SIMD_MODE
-
